Move methods from SolverConfigs to SolverConfig
