+++ optaplanner-core/src/main/java/org/optaplanner/core/config/solver/SolverConfig.java
@@ -16,13 +16,24 @@
 
 package org.optaplanner.core.config.solver;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.Reader;
+import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.ThreadFactory;
 
+import com.thoughtworks.xstream.XStream;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
 import com.thoughtworks.xstream.annotations.XStreamImplicit;
+import com.thoughtworks.xstream.converters.ConversionException;
 import org.optaplanner.core.api.solver.Solver;
 import org.optaplanner.core.api.solver.SolverFactory;
 import org.optaplanner.core.config.AbstractConfig;
@@ -42,6 +53,7 @@ import org.optaplanner.core.impl.phase.Phase;
 import org.optaplanner.core.impl.score.definition.ScoreDefinition;
 import org.optaplanner.core.impl.score.director.InnerScoreDirectorFactory;
 import org.optaplanner.core.impl.solver.DefaultSolver;
+import org.optaplanner.core.impl.solver.io.XStreamConfigReader;
 import org.optaplanner.core.impl.solver.random.DefaultRandomFactory;
 import org.optaplanner.core.impl.solver.random.RandomFactory;
 import org.optaplanner.core.impl.solver.recaller.BestSolutionRecaller;
@@ -54,11 +66,133 @@ import org.slf4j.LoggerFactory;
 import static org.apache.commons.lang3.ObjectUtils.*;
 
 /**
- * To read from XML, use {@link SolverConfigs#createFromXmlResource(String)}.
+ * To read from XML, use {@link #createFromXmlResource(String)}.
  * To build a {@link SolverFactory}, use {@link SolverFactory#create(SolverConfig)}.
  */
 @XStreamAlias("solver")
 public class SolverConfig extends AbstractConfig<SolverConfig> {
+
+    /**
+     * Reads an XML solver configuration from the classpath.
+     * @param solverConfigResource never null, a classpath resource
+     * as defined by {@link ClassLoader#getResource(String)}
+     * @return never null
+     */
+    public static SolverConfig createFromXmlResource(String solverConfigResource) {
+        return createFromXmlResource(solverConfigResource, null);
+    }
+
+    /**
+     * See {@link #createFromXmlResource(String)}.
+     * @param solverConfigResource never null, a classpath resource
+     * as defined by {@link ClassLoader#getResource(String)}
+     * @param classLoader sometimes null, the {@link ClassLoader} to use for loading all resources and {@link Class}es,
+     *      null to use the default {@link ClassLoader}
+     * @return never null
+     */
+    public static SolverConfig createFromXmlResource(String solverConfigResource, ClassLoader classLoader) {
+        ClassLoader actualClassLoader = classLoader != null ? classLoader : SolverConfig.class.getClassLoader();
+        try (InputStream in = actualClassLoader.getResourceAsStream(solverConfigResource)) {
+            if (in == null) {
+                String errorMessage = "The solverConfigResource (" + solverConfigResource
+                        + ") does not exist as a classpath resource in the classLoader (" + actualClassLoader + ").";
+                if (solverConfigResource.startsWith("/")) {
+                    errorMessage += "\nA classpath resource should not start with a slash (/)."
+                            + " A solverConfigResource adheres to ClassLoader.getResource(String)."
+                            + " Maybe remove the leading slash from the solverConfigResource.";
+                }
+                throw new IllegalArgumentException(errorMessage);
+            }
+            return createFromXmlInputStream(in, classLoader);
+        } catch (ConversionException e) {
+            String lineNumber = e.get("line number");
+            throw new IllegalArgumentException("Unmarshalling of solverConfigResource (" + solverConfigResource
+                    + ") fails on line number (" + lineNumber + ")."
+                    + (Objects.equals(e.get("required-type"), "java.lang.Class")
+                    ? "\n  Maybe the classname on line number (" + lineNumber + ") is surrounded by whitespace, which is invalid."
+                    : ""), e);
+        } catch (IOException e) {
+            throw new IllegalArgumentException("Reading the solverConfigResource (" + solverConfigResource + ") failed.", e);
+        }
+    }
+
+    /**
+     * Reads an XML solver configuration from the file system.
+     * <p>
+     * Warning: this leads to platform dependent code,
+     * it's recommend to use {@link #createFromXmlResource(String)} instead.
+     * @param solverConfigFile never null
+     * @return never null
+     */
+    public static SolverConfig createFromXmlFile(File solverConfigFile) {
+        return createFromXmlFile(solverConfigFile, null);
+    }
+
+    /**
+     * See {@link #createFromXmlFile(File)}.
+     * @param solverConfigFile never null
+     * @param classLoader sometimes null, the {@link ClassLoader} to use for loading all resources and {@link Class}es,
+     *      null to use the default {@link ClassLoader}
+     * @return never null
+     */
+    public static SolverConfig createFromXmlFile(File solverConfigFile, ClassLoader classLoader) {
+        try (InputStream in = new FileInputStream(solverConfigFile)) {
+            return createFromXmlInputStream(in, classLoader);
+        } catch (FileNotFoundException e) {
+            throw new IllegalArgumentException("The solverConfigFile (" + solverConfigFile + ") was not found.", e);
+        } catch (IOException e) {
+            throw new IllegalArgumentException("Reading the solverConfigFile (" + solverConfigFile + ") failed.", e);
+        }
+    }
+
+    /**
+     * @param in never null, gets closed
+     * @return never null
+     */
+    public static SolverConfig createFromXmlInputStream(InputStream in) {
+        return createFromXmlInputStream(in, null);
+    }
+
+    /**
+     * See {@link #createFromXmlInputStream(InputStream)}.
+     * @param in never null, gets closed
+     * @param classLoader sometimes null, the {@link ClassLoader} to use for loading all resources and {@link Class}es,
+     *      null to use the default {@link ClassLoader}
+     * @return never null
+     */
+    public static SolverConfig createFromXmlInputStream(InputStream in, ClassLoader classLoader) {
+        try (Reader reader = new InputStreamReader(in, "UTF-8")) {
+            return createFromXmlReader(reader, classLoader);
+        } catch (UnsupportedEncodingException e) {
+            throw new IllegalStateException("This vm does not support UTF-8 encoding.", e);
+        } catch (IOException e) {
+            throw new IllegalArgumentException("Reading solverConfigInputStream failed.", e);
+        }
+    }
+
+    /**
+     * @param reader never null, gets closed
+     * @return never null
+     */
+    public static SolverConfig createFromXmlReader(Reader reader) {
+        return createFromXmlReader(reader, null);
+    }
+
+    /**
+     * See {@link #createFromXmlReader(Reader)}.
+     * @param reader never null, gets closed
+     * @param classLoader sometimes null, the {@link ClassLoader} to use for loading all resources and {@link Class}es,
+     *      null to use the default {@link ClassLoader}
+     * @return never null
+     */
+    public static SolverConfig createFromXmlReader(Reader reader, ClassLoader classLoader) {
+        XStream xStream = XStreamConfigReader.buildXStream(classLoader);
+        return (SolverConfig) xStream.fromXML(reader);
+    }
+
+    // ************************************************************************
+    // Fields
+    // ************************************************************************
 
     public static final String MOVE_THREAD_COUNT_NONE = "NONE";
     public static final String MOVE_THREAD_COUNT_AUTO = "AUTO";
