+++ optaplanner-core/src/test/java/org/optaplanner/core/impl/solver/XStreamXmlSolverFactoryTest.java
@@ -24,7 +24,6 @@ import org.apache.commons.io.IOUtils;
 import org.junit.Test;
 import org.optaplanner.core.config.SolverConfigContext;
 import org.optaplanner.core.config.solver.SolverConfig;
-import org.optaplanner.core.config.solver.SolverConfigs;
 import org.optaplanner.core.impl.solver.io.XStreamConfigReader;
 
 import static org.junit.Assert.*;
@@ -36,7 +35,7 @@ public class XStreamXmlSolverFactoryTest {
         String solverConfigResource = "testdataSolverConfigXStream.xml";
         String originalXml = IOUtils.toString(getClass().getResourceAsStream(solverConfigResource), "UTF-8");
         InputStream originalConfigInputStream = getClass().getResourceAsStream(solverConfigResource);
-        SolverConfig solverConfig = SolverConfigs.createFromXmlInputStream(originalConfigInputStream);
+        SolverConfig solverConfig = SolverConfig.createFromXmlInputStream(originalConfigInputStream);
         SolverConfigContext configContext = new SolverConfigContext(getClass().getClassLoader());
         assertNotNull(solverConfig.buildSolver(configContext));
         XStream xStream = XStreamConfigReader.buildXStream(getClass().getClassLoader());
