+++ optaplanner-examples/src/test/java/org/optaplanner/examples/common/app/RealTimePlanningTurtleTest.java
@@ -27,7 +27,6 @@ import org.optaplanner.core.api.domain.solution.PlanningSolution;
 import org.optaplanner.core.api.solver.Solver;
 import org.optaplanner.core.api.solver.SolverFactory;
 import org.optaplanner.core.config.solver.SolverConfig;
-import org.optaplanner.core.config.solver.SolverConfigs;
 import org.optaplanner.core.config.solver.termination.TerminationConfig;
 import org.optaplanner.core.impl.solver.ProblemFactChange;
 
@@ -54,7 +53,7 @@ public abstract class RealTimePlanningTurtleTest<Solution_> extends AbstractTurt
     }
 
     protected SolverFactory<Solution_> buildSolverFactory() {
-        SolverConfig solverConfig = SolverConfigs.createFromXmlResource(createSolverConfigResource());
+        SolverConfig solverConfig = SolverConfig.createFromXmlResource(createSolverConfigResource());
         solverConfig.setDaemon(true);
         solverConfig.setTerminationConfig(new TerminationConfig().withMillisecondsSpentLimit(SPENT_LIMIT));
         return SolverFactory.create(solverConfig);
